Added Singleton and Timer To Chest using System.DateTime

+++ Assets/Scripts/ChestScriptableObjectList.cs
@@ -15,5 +15,5 @@ public class ChestScriptableObject
     public int maxCoins;
     public int minGems;
     public int maxGems;
-    public float TimeToUnlockInSeconds;
+    public int TimeToUnlockInSeconds;
 }
